Fix warnings in ObjectAL

+++ src/platform/ios/external/objectAL/Support/OALTools.m
@@ -113,7 +113,7 @@
 		va_start(args, description);
 		description = [[NSString alloc] initWithFormat:description arguments:args];
 		va_end(args);
-		OAL_LOG_ERROR_CONTEXT(function, @"%@ (error code 0x%08x: %@)", description, errorCode, errorString);
+		OAL_LOG_ERROR_CONTEXT(function, @"%@ (error code 0x%08x: %@)", description, (unsigned int)errorCode, errorString);
 		arcsafe_release(description);
 	}
 }
